Refactoring of table
