+++ libs/pxweb2-ui/src/index.ts
@@ -10,6 +10,7 @@ export * from './lib/components/Typography/BodyLong/BodyLong';
 export * from './lib/components/Icon/Icon';
 export * from './lib/components/Search/Search';
 export * from './lib/components/Table/Table';
+export * from './lib/components/Table/tableBuilder';
 export * from './lib/shared-types/codelist';
 export * from './lib/shared-types/codelistType';
 export * from './lib/shared-types/note';
@@ -19,4 +20,4 @@ export * from './lib/shared-types/pxTableData';
 export * from './lib/shared-types/value';
 export * from './lib/shared-types/variable';
 export * from './lib/shared-types/vartypeEnum';
-export * from './lib/shared-types/multiDimensionalData';
+
